specify ", advanced edition" and ", standard edition" of BWZ where appropriate
I'm not sure if all of these are correct, I cannot verify as I do not have an advance edition subscription, but I think for the most part this tutorial is only appropriate for advanced edition. As far as I can tell some of the prerequisites are also only for advanced edition, in particular because the `Workzone_Admin` role is not available in standard edition (I think).